Avoid code in an __init__ module
